server/api/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0
httpx==0.26.0
supabase==2.3.1
python-jose==3.3.0
cryptography==41.0.7
python-multipart==0.0.9
openai==1.10.0
numpy==1.26.2
tenacity==8.2.3